Avoid modifying the private state of the inventory
Possiby due to https://github.com/PrismarineJS/prismarine-windows/pull/41 mineflayer-web-inventory was modifying the private state of the mineflayer bot's inventory which was causing strange errors to happen.
Apparently the error didn't happen when using moveSlotItem function, so the tests didn't catch it. Added another test for the deposit function. TODO: another test for the withdraw function.

+ general improvements

Should close #19 and #20